+++ test/test_ecc_base.jl
@@ -2,14 +2,19 @@ using Test
 using QuantumClifford.ECC.QECCore
 using QuantumClifford
 using QuantumClifford.ECC
-using QuantumClifford.ECC: check_repr_commutation_relation
+using QuantumClifford.ECC: check_repr_commutation_relation, check_repr_regular_linear
 using InteractiveUtils
+using SparseArrays
 
 import Nemo: GF
 import LinearAlgebra
-import Hecke: group_algebra, abelian_group, gens, quo, one
+import Hecke: group_algebra, abelian_group, gens, quo, one, small_group
 
 # generate instances of all implemented codes to make sure nothing skips being checked
+
+const H1 = sparse(Bool[1 0 1 0; 0 1 0 1; 1 1 0 0]);
+
+const H2 = sparse(Bool[1 1 0;0 1 1]);
 
 # We do not include smaller random circuit code because some of them has a bad distance and fails the TableDecoder test
 const random_brickwork_circuit_args = repeat([((20,), 50, [1]), ((20,), 50, 1:2:20), ((5, 5), 50, [1]), ((3, 3, 3), 50, [1])], 10)
@@ -44,8 +49,41 @@ test_gb_codes = [
 ]
 
 test_hcubic_codes = [
-    haah_cubic_codes([0, 15, 20, 28, 66], [0, 58, 59, 100, 121], 3)
+    haah_cubic_codes([0, 15, 20, 28, 66], [0, 58, 59, 100, 121], 3),
+    haah_cubic_codes(8), # (D) [[1024, 30, 13 ≤ d ≤ 32]] Appendix B of [panteleev2021degenerate](@cite).
 ]
+
+# honeycomb color codes from [eberhardt2024logical](@cite).
+test_honeycomb_color_codes = [
+    honeycomb_color_codes(6 , 6), honeycomb_color_codes(9 , 6),
+    honeycomb_color_codes(12, 6), honeycomb_color_codes(12, 9),
+]
+
+# Lifted product codes using non-commutative algebras
+# These are just made up, they are not known to
+# specifically be good, they are not from a paper.
+G = small_group(36,1)
+GA = group_algebra(GF(2), G)
+r, s  = gens(GA);
+A = 1 + r
+B = 1 + s + r^6 + s^3*r + s*r^7 + s^3*r^5
+nonabel1 = two_block_group_algebra_codes(A,B)
+
+G = small_group(48,10)
+GA = group_algebra(GF(2), G)
+r, s  = gens(GA);
+A = 1 + s*r^2
+B = 1 + r + s^3 + s^4 + s^2*r^5 + s^4*r^6
+nonabel2 = two_block_group_algebra_codes(A,B)
+
+G = small_group(40,8)
+GA = group_algebra(GF(2), G)
+r, s  = gens(GA);
+A = 1 + s*r^5 + r^5 + s*r^6
+B = 1 + s^2 + r + s^2*r^3
+nonabel3 = two_block_group_algebra_codes(A,B)
+
+test_nonabelian_codes = [nonabel1, nonabel2, nonabel3]
 
 other_lifted_product_codes = []
 
@@ -53,6 +91,7 @@ other_lifted_product_codes = []
 l = 63
 GA = group_algebra(GF(2), abelian_group(l))
 @test check_repr_commutation_relation(GA) # TODO use this check more pervasively throughout the test suite
+@test check_repr_regular_linear(GA) # TODO use this check more pervasively throughout the test suite
 A = zeros(GA, 7, 7)
 x = gens(GA)[]
 A[LinearAlgebra.diagind(A)] .= x^27
@@ -153,7 +192,7 @@ test_bb_codes = [bb1, bb2, bb3]
 test_twobga_codes = []
 
 @static if !Sys.iswindows() && Sys.ARCH == :x86_64 && VERSION >= v"1.11"
-  import Oscar: free_group, cyclic_group, direct_product
+  import Oscar: free_group, cyclic_group, direct_product, small_group_identification, describe, order, gens, quo
   function load_oscar_codes()
     #@info "Add group theoretic codes requiring Oscar"
     # [[72, 8, 9]] 2BGA code taken from Table I Block 1 of [lin2024quantum](@cite)
@@ -219,7 +258,50 @@ test_twobga_codes = []
     b = [one(GA), x^3, s * x^6, x^4, s * x^9, s * x^10]
     dprod2 = twobga_from_direct_product(a, b, GA)
 
-    append!(test_twobga_codes, [t1b1, t1b3, tb21, tb22, dprod1, dprod2])
+    # 2BGA codes using non-abelian groups: Table III of [lin2024quantum](@cite)
+    # [[24, 8, 3]]
+    m = 6
+    F = free_group(["r", "s"])
+    r, s = gens(F)
+    G, = quo(F, [r^m, s^2, (r*s)^2])
+    GA = group_algebra(GF(2), G)
+    r, s = gens(G)
+    a = [one(G), r^4]
+    b = [one(G), s*r^4, r^3, r^4, s*r^2, r]
+    nonabel1 = twobga_from_fp_group(a, b, GA)
+
+    # [[24, 12, 2]]
+    F = free_group(["r", "s"])
+    r, s = gens(F)
+    G, = quo(F, [r^m, s^2, (r*s)^2])
+    GA = group_algebra(GF(2), G)
+    r, s = gens(G)
+    a = [one(G), r^3]
+    b = [one(G), s*r, r^3, r^4, s*r^4, r]
+    nonabel2 = twobga_from_fp_group(a, b, GA)
+
+    # [[32, 8, 4]]
+    m = 8
+    F = free_group(["r", "s"])
+    r, s = gens(F)
+    G, = quo(F, [r^m, s^2, (r*s)^2])
+    GA = group_algebra(GF(2), G)
+    r, s = gens(G)
+    a = [one(G), r^2]
+    b = [one(G), s*r^5, s*r^4, r^2, s*r^7, s*r^6]
+    nonabel3 = twobga_from_fp_group(a, b, GA)
+
+    # [[32, 16, 2]]
+    F = free_group(["r", "s"])
+    r, s = gens(F)
+    G, = quo(F, [r^m, s^2, (r*s)^2])
+    GA = group_algebra(GF(2), G)
+    r, s = gens(G)
+    a = [one(G), r^4]
+    b = [one(G), s*r^3, s*r^6, r^4, s*r^7, s*r^2]
+    nonabel4 = twobga_from_fp_group(a, b, GA)
+
+    append!(test_twobga_codes, [t1b1, t1b3, tb21, tb22, dprod1, dprod2, nonabel1, nonabel2, nonabel3, nonabel4])
   end
   load_oscar_codes()
 end
@@ -233,10 +315,15 @@ const code_instance_args = Dict(
     :Concat => [(Perfect5(), Perfect5()), (Perfect5(), Steane7()), (Steane7(), Cleve8()), (Toric(2, 2), Shor9())],
     :CircuitCode => random_circuit_code_args,
     :QuantumReedMuller => [3, 4, 5],
-    :LPCode => (c -> (c.A, c.B)).(vcat(LP04, LP118, test_gb_codes, test_bb_codes, test_mbb_codes, test_coprimeBB_codes, test_hcubic_codes, test_twobga_codes, other_lifted_product_codes)),
+    :LPCode => (c -> (c.A, c.B)).(vcat(LP04, LP118, test_gb_codes, test_bb_codes, test_mbb_codes, test_coprimeBB_codes, test_hcubic_codes, test_twobga_codes, test_honeycomb_color_codes, test_nonabelian_codes, other_lifted_product_codes)),
     :QuantumReedMuller => [3, 4, 5],
     :Triangular488 => [3, 5, 7, 9, 11],
-    :Triangular666 => [3, 5, 7, 9, 11]
+    :Triangular666 => [3, 5, 7, 9, 11],
+    :DelfosseReichardt => [(2,1,3), (2,2,4), (4,3,5), (4,3,6)],
+    :DelfosseReichardtRepCode => [4, 6, 8, 10],
+    :DelfosseReichardt823 => [1, 2, 3, 4, 5],
+    :QuantumTannerGraphProduct => [(H1, H2),(H2, H2), (H1, H1), (H2, H1)],
+    :CyclicQuantumTannerGraphProduct => [1, 2, 3, 4, 5]
 )
 
 function all_testablable_code_instances(;maxn=nothing)