+++ test/test_noisycircuits.jl
@@ -295,4 +295,35 @@
             @test stabilizerview(r) == expectedFinalState
         end
     end
+
+    @testset "VerifyOp" begin
+        using QuantumClifford.ECC: Steane7, parity_checks, naive_encoding_circuit
+        @testset "Stabilizer passed as good_state is not a logical state" begin
+            good_state = parity_checks(Steane7()) #passing in a code instead of a state within codespace
+            verify = VerifyOp(good_state, 1:7)
+            reg = Register(one(MixedDestabilizer,7),6) #dummy register to pass into applywstatus!
+        
+            
+            @test_throws ArgumentError applywstatus!(reg, verify) #should throw an error since good_state isn't a logical state
+            
+    
+        end
+    
+        @testset "Accepts pure good_state argument" begin
+            good_state = S"ZZI
+                            IZZ 
+                            XXX" # passing in GHZ state as good_state argument since it exists within the repetition code's codespace
+            verify = VerifyOp(good_state, 1:3)
+            reg = Register(one(MixedDestabilizer,3),2) 
+            encoding = [sHadamard(1)  ,  sCNOT(1,2)   ,sCNOT(1,3) ] # encoding the register's state into a logical state within the repetition code's codespace
+            for i in encoding
+                apply!(reg, i)
+            end
+    
+            _, status = applywstatus!(reg, verify)
+            @test status == true_success_stat
+        end
+    
+      
+    end
 end